+++ NEWS.md
@@ -53,6 +53,15 @@ that simplifies the use of an external .hpp file to compile with the model.
 
 * Added E-BFMI checks that run automatically post sampling. (#500, @jsocolar)
 
+* New method `$code()` for all fitted model objects that returns the Stan code
+associated with the fitted model. (#575)
+
+* Added the `cmdstanr_force_recompile` global option that is used for forcing
+recompilation of Stan models. (#580)
+
+* New methods for `posterior::as_draws()` for CmdStanR fitted model objects.
+These are just wrappers around the `$draws()` method provided for convenience. (#532)
+
 # cmdstanr 0.4.0
 
 ### Bug fixes